Cleaned up typography
