+++ src/cassandra_ast/create_table.rs
@@ -1,6 +1,6 @@
-use std::fmt::{Display, Formatter};
-use itertools::Itertools;
 use crate::cassandra_ast::common::{ColumnDefinition, PrimaryKey, WithItem};
+use itertools::Itertools;
+use std::fmt::{Display, Formatter};
 
 /// The data for a `Create table` statement
 #[derive(PartialEq, Debug, Clone)]
