+++ src/cassandra_ast/update.rs
@@ -1,8 +1,8 @@
-use std::fmt::{Display, Formatter};
-use itertools::Itertools;
 use crate::cassandra_ast::begin_batch::BeginBatch;
 use crate::cassandra_ast::common::{Operand, RelationElement, TtlTimestamp};
 use crate::cassandra_ast::delete::IndexedColumn;
+use itertools::Itertools;
+use std::fmt::{Display, Formatter};
 
 /// data for `Update` statements
 #[derive(PartialEq, Debug, Clone)]
